pyproject.toml
[project]
name = "stochastic-rsi-backtesting"
version = "0.1.0"
description = "Backtesting framework for a MA200 + Stochastic RSI strategy with Streamlit UI."
readme = "Readme.md"
requires-python = ">=3.10,<3.13"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
  "backtesting==0.3.3",
  "matplotlib==3.8.2",
  "numpy==1.26.2",
  "pandas==2.1.3",
  "plotly==5.18.0",
  "psycopg2-binary==2.9.9",
  "python-dateutil==2.8.2",
  "python-dotenv==1.0.0",
  "pytz==2023.3",
  "sqlalchemy==2.0.23",
  "streamlit==1.29.0"
]

[project.optional-dependencies]
dev = [
  "black==23.12.0",
  "flake8==6.1.0",
  "pytest==7.4.3",
  "pytest-cov==4.1.0",
  "watchdog==4.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
